+++ package.json
@@ -48,6 +48,7 @@
     "async-mutex": "^0.3.2",
     "bignumber.js": "^9.0.2",
     "cache-manager": "^3.6.0",
+    "cbor": "^8.1.0",
     "class-transformer": "^0.5.1",
     "class-validator": "^0.13.2",
     "defichain": "^2.15.0",
@@ -57,7 +58,6 @@
     "leveldown": "6.0.0",
     "lexicographic-integer-encoding": "^1.0.1",
     "memdown": "^6.1.1",
-    "msgpack-lite": "^0.1.26",
     "reflect-metadata": "^0.1.13",
     "rimraf": "^3.0.2",
     "rxjs": "^7.5.2",
@@ -73,7 +73,6 @@
     "@types/level": "^6.0.0",
     "@types/lossless-json": "^1.0.1",
     "@types/memdown": "^3.0.0",
-    "@types/msgpack-lite": "^0.1.8",
     "@types/node": "^14.14.36",
     "@types/subleveldown": "^4.1.1",
     "eslint": "^7.32.0",
